backend/ai/retrievers/scholar_retriever.py
"""
Google Scholar Retriever for NeuroQuest

Note: Google Scholar does not provide an official API. This implementation
uses alternative approaches and should be used in compliance with Google's
Terms of Service.
"""
import requests
from typing import List, Dict, Any, Optional
from bs4 import BeautifulSoup
import time
import re


class ScholarRetriever:
    """Retriever for Google Scholar papers."""

    BASE_URL = "https://scholar.google.com"

    def __init__(self, max_results: int = 10):
        """
        Initialize the Google Scholar retriever.

        Args:
            max_results: Maximum number of results to retrieve
        """
        self.max_results = max_results
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        })

    def search(self, query: str, max_results: Optional[int] = None) -> List[Dict[str, Any]]:
        """
        Search Google Scholar for papers.

        Args:
            query: Search query
            max_results: Maximum number of results (overrides instance default)

        Returns:
            List of search results

        Note: This is a simplified implementation. For production use,
        consider using official APIs or alternative services.
        """
        max_results = max_results or self.max_results

        # For MVP purposes, return mock results
        # In production, implement proper web scraping or use alternative APIs
        return self._get_mock_results(query, max_results)

    def _get_mock_results(self, query: str, max_results: int) -> List[Dict[str, Any]]:
        """
        Get mock results for development purposes.

        Args:
            query: Search query
            max_results: Maximum number of results

        Returns:
            List of mock search results
        """
        mock_results = []

        for i in range(max_results):
            mock_results.append({
                'id': f"scholar_{i}",
                'title': f"Google Scholar Paper {i+1}: {query}",
                'authors': [f"Author {j+1}" for j in range(3)],
                'abstract': f"This is a mock abstract for Google Scholar paper {i+1} about {query}. "
                           f"In production, this would be fetched from Google Scholar.",
                'source': 'scholar',
                'url': f"https://scholar.google.com/scholar?q={query.replace(' ', '+')}",
                'published_date': "2024-01-01",
                'relevance_score': 0.85 - (i * 0.05),
                'citation_count': 100 - i * 10,
                'metadata': {
                    'venue': 'Conference/Journal',
                    'year': 2024,
                }
            })

        return mock_results

    def _search_scholar_real(self, query: str, max_results: int) -> List[Dict[str, Any]]:
        """
        Real Google Scholar search implementation (not used in MVP).

        Args:
            query: Search query
            max_results: Maximum number of results

        Returns:
            List of search results

        Warning: This method may violate Google's Terms of Service.
        Use at your own risk and consider using official APIs.
        """
        params = {
            'q': query,
            'start': 0,
            'num': max_results,
        }

        try:
            response = self.session.get(
                f"{self.BASE_URL}/scholar",
                params=params,
                timeout=10
            )
            response.raise_for_status()

            soup = BeautifulSoup(response.content, 'html.parser')

            results = []
            for div in soup.find_all('div', class_='gs_ri'):
                result = self._parse_scholar_result(div)
                if result:
                    results.append(result)

            return results

        except Exception as e:
            print(f"Error searching Google Scholar: {e}")
            return []

    def _parse_scholar_result(self, div: Any) -> Optional[Dict[str, Any]]:
        """
        Parse a Google Scholar result div.

        Args:
            div: BeautifulSoup element

        Returns:
            Parsed result dictionary
        """
        try:
            # Extract title and URL
            title_link = div.find('h3', class_='gs_rt').find('a')
            title = title_link.text if title_link else ''
            url = title_link.get('href', '') if title_link else ''

            # Extract authors and venue
            author_info = div.find('div', class_='gs_a')
            authors = []
            venue = ''
            year = ''

            if author_info:
                text = author_info.text
                # Parse authors (usually before the dash)
                parts = text.split(' - ')
                if parts:
                    authors_text = parts[0]
                    authors = [a.strip() for a in authors_text.split(',')]

                    # Extract year from venue info
                    if len(parts) > 1:
                        venue_info = parts[1]
                        year_match = re.search(r'\d{4}', venue_info)
                        if year_match:
                            year = year_match.group()

            # Extract abstract
            abstract_div = div.find('div', class_='gs_rs')
            abstract = abstract_div.text if abstract_div else ''

            # Extract citation count
            citation_link = div.find('a', string=re.compile(r'Cited by'))
            citation_count = 0
            if citation_link:
                citation_match = re.search(r'Cited by (\d+)', citation_link.text)
                if citation_match:
                    citation_count = int(citation_match.group(1))

            return {
                'id': f"scholar_{hash(url)}",
                'title': title,
                'authors': authors,
                'abstract': abstract,
                'source': 'scholar',
                'url': url,
                'published_date': f"{year}-01-01" if year else None,
                'relevance_score': 0.8,
                'citation_count': citation_count,
                'metadata': {
                    'venue': venue,
                    'year': year,
                }
            }
        except Exception as e:
            print(f"Error parsing Google Scholar result: {e}")
            return None


# Alternative: Use Semantic Scholar API (recommended for production)
class SemanticScholarRetriever:
    """Retriever using Semantic Scholar API (recommended alternative)."""

    BASE_URL = "https://api.semanticscholar.org/graph/v1"

    def __init__(self, api_key: Optional[str] = None, max_results: int = 10):
        """
        Initialize the Semantic Scholar retriever.

        Args:
            api_key: Semantic Scholar API key (optional)
            max_results: Maximum number of results to retrieve
        """
        self.api_key = api_key
        self.max_results = max_results
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'NeuroQuest/1.0 (https://github.com/neuroquest)'
        })

        if api_key:
            self.session.headers.update({'x-api-key': api_key})

    def search(self, query: str, max_results: Optional[int] = None) -> List[Dict[str, Any]]:
        """
        Search Semantic Scholar for papers.

        Args:
            query: Search query
            max_results: Maximum number of results

        Returns:
            List of search results
        """
        max_results = max_results or self.max_results

        params = {
            'query': query,
            'limit': max_results,
            'fields': 'paperId,title,abstract,authors,year,citationCount,url,venue'
        }

        try:
            response = self.session.get(
                f"{self.BASE_URL}/paper/search",
                params=params,
                timeout=10
            )
            response.raise_for_status()

            data = response.json()
            papers = data.get('data', [])

            results = []
            for paper in papers:
                result = self._parse_paper(paper)
                if result:
                    results.append(result)

            return results

        except Exception as e:
            print(f"Error searching Semantic Scholar: {e}")
            return []

    def _parse_paper(self, paper: Dict[str, Any]) -> Optional[Dict[str, Any]]:
        """
        Parse a Semantic Scholar paper.

        Args:
            paper: Paper data from API

        Returns:
            Parsed result dictionary
        """
        try:
            authors = [author.get('name', '') for author in paper.get('authors', [])]

            return {
                'id': f"semantic_{paper.get('paperId', '')}",
                'title': paper.get('title', ''),
                'authors': authors,
                'abstract': paper.get('abstract', ''),
                'source': 'semantic_scholar',
                'url': paper.get('url', ''),
                'published_date': f"{paper.get('year', '')}-01-01" if paper.get('year') else None,
                'relevance_score': 0.8,
                'citation_count': paper.get('citationCount', 0),
                'metadata': {
                    'venue': paper.get('venue', ''),
                    'year': paper.get('year', ''),
                }
            }
        except Exception as e:
            print(f"Error parsing Semantic Scholar paper: {e}")
            return None


# Example usage
if __name__ == "__main__":
    # Use mock retriever for MVP
    retriever = ScholarRetriever(max_results=5)

    results = retriever.search("machine learning")
    print(f"Found {len(results)} results")

    for result in results[:2]:
        print(f"\nTitle: {result['title']}")
        print(f"Authors: {', '.join(result['authors'])}")
        print(f"URL: {result['url']}")

    # For production, consider using Semantic Scholar API
    # semantic_retriever = SemanticScholarRetriever(max_results=5)
    # semantic_results = semantic_retriever.search("machine learning")
