backend/ai/requirements.txt
# AI/ML Dependencies
langchain==0.1.0
langchain-community==0.0.10
anthropic==0.7.8
openai==1.3.7
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# NLP
spacy==3.7.2

# Data Processing
numpy==1.26.2
pandas==2.1.4

# HTTP
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.2

# Graph Processing
networkx==3.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
